(maint) Update nuspecs to remove templated values
This commit is an initial pass at updating all of the nuspec files used
for tests to pass the Chocolatey Community Repository's validation
rules. We suggest guidelines for packages, so our test packages should
meet those guidelines. The update of the description also allows us to
better see the purpose of packages in a NuGet repository without needing
the source package.