(tests) Update upgrade all test to exclude choco
Sometimes tests in Test Kitchen are run against versions of Chocolatey
that are available on the setup repository, and so they are found as
part of upgrade testing. The tests don't need these newer Chocolatey
versions, nor do we expect to find them. This excludes all of the
packages that are known to be on testing environments but not part of
the test in question.

This is to reduce the number of false positive tests we have, especially
as we come up to a release.