(tests) Run Get-EnvironmentVariableNames tests
The $variables passed to the ForEach for these tests was found to be
null, and so these tests weren't actually running. Upon looking at the
tests, determined the ForEach variable was unintentional and was
removed.

The tests get the variable names in one order, and the .Net method to
get the names presents them in a different order. Changed this to sort
the list for both of them so that the Pester tests were consistent.