(tests) Use Chocolatey Test Environment for tests
We use the Chocolatey Test Environment for testing packages on the
Chocolatey Community Repository, so we should probably use it here for
testing Chocolatey itself.

This also adds a copy of the `testResults.xml` generated by Pester from
the test system to the host system so it can be analyzed.